requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bcrypt==3.1.7
black==19.10b0
bleach==3.1.0
boto3==1.12.4
botocore==1.15.4
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
chart-studio==1.0.0
Click==7.0
colorama==0.4.1
cryptography==2.7
cycler==0.10.0
dash==1.8.0
dash-canvas==0.1.0
dash-core-components==1.7.0
dash-html-components==1.0.2
dash-renderer==1.2.3
dash-table==4.6.0
decorator==4.4.0
defusedxml==0.6.0
distlib==0.3.0
docutils==0.15.2
entrypoints==0.3
filelock==3.0.12
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
gevent==1.4.0
graphviz==0.13
greenlet==0.4.15
gunicorn==20.0.4
idna==2.8
imageio==2.8.0
importlib-metadata==0.18
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.20
itsdangerous==1.1.0
jedi==0.14.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.14.0
Js2Py==0.67
json5==0.8.5
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.1.4
jupyterlab-black==0.2.1
jupyterlab-code-formatter==0.2.3
jupyterlab-server==1.0.6
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
line-profiler==2.1.2
llvmlite==0.30.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
memory-profiler==0.55.0
mistune==0.8.4
more-itertools==7.2.0
nb-black==1.0.6
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.4
notebook==5.7.8
numba==0.46.0
numpy==1.17.3
objgraph==3.4.1
pandas==0.25.2
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.5.0
pathspec==0.7.0
pi==0.1.2
pickleshare==0.7.5
Pillow==7.0.0
plotly==4.2.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psutil==5.6.3
pycparser==2.19
Pygments==2.4.2
pyjsparser==2.7.1
pylint==2.4.4
PyNaCl==1.3.0
pyparsing==2.4.2
pyrsistent==0.15.2
pysftp==0.2.9
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.1.1
pywin32==225
pywinpty==0.5.5
pyzmq==18.0.1
qtconsole==4.5.5
regex==2020.2.20
requests==2.22.0
retrying==1.3.3
s3transfer==0.3.3
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
simplejson==3.16.0
six==1.12.0
sklearn==0.0
terminado==0.8.2
testpath==0.4.2
toml==0.10.0
tornado==6.0.3
traitlets==4.3.2
typed-ast==1.4.0
tzlocal==2.0.0
urllib3==1.25.3
virtualenv==20.0.8
virtualenvwrapper-win==1.2.6
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.5.1
